Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
requests==2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
APScheduler==3.10.4
gunicorn==21.2.0
Werkzeug==2.3.7
redis==5.0.1
boto3==1.34.0
prometheus-client==0.19.0
beautifulsoup4==4.12.2
lxml==4.9.3
h5py==3.10.0
matplotlib==3.8.2
Pillow==10.1.0
scipy==1.11.4
